Add refresh to domain

+++ domain/record.go
@@ -10,19 +10,25 @@ import (
 	"github.com/pritunl/mongo-go-driver/mongo/options"
 	"github.com/pritunl/pritunl-cloud/database"
 	"github.com/pritunl/pritunl-cloud/errortypes"
+	"github.com/pritunl/pritunl-cloud/settings"
 	"github.com/pritunl/pritunl-cloud/utils"
 )
 
 type Record struct {
-	Id         primitive.ObjectID `bson:"_id,omitempty" json:"id"`
-	Domain     primitive.ObjectID `bson:"domain" json:"domain"`
-	Resource   primitive.ObjectID `bson:"resource,omitempty" json:"resource"`
-	Deployment primitive.ObjectID `bson:"deployment,omitempty" json:"deployment"`
-	Timestamp  time.Time          `bson:"timestamp" json:"timestamp"`
-	SubDomain  string             `bson:"sub_domain" json:"sub_domain"`
-	Type       string             `bson:"type" json:"type"`
-	Value      string             `bson:"value" json:"value"`
-	Operation  string             `bson:"-" json:"operation"`
+	Id              primitive.ObjectID `bson:"_id,omitempty" json:"id"`
+	Domain          primitive.ObjectID `bson:"domain" json:"domain"`
+	Resource        primitive.ObjectID `bson:"resource,omitempty" json:"resource"`
+	Deployment      primitive.ObjectID `bson:"deployment,omitempty" json:"deployment"`
+	Timestamp       time.Time          `bson:"timestamp" json:"timestamp"`
+	DeleteTimestamp time.Time          `bson:"delete_timestamp,omitempty" json:"delete_timestamp"`
+	SubDomain       string             `bson:"sub_domain" json:"sub_domain"`
+	Type            string             `bson:"type" json:"type"`
+	Value           string             `bson:"value" json:"value"`
+	Operation       string             `bson:"-" json:"operation"`
+}
+
+func (r *Record) IsDeleted() bool {
+	return !r.DeleteTimestamp.IsZero()
 }
 
 func (r *Record) Copy() *Record {
@@ -92,6 +98,22 @@ func (r *Record) CommitFields(db *database.Database, fields set.Set) (
 
 func (r *Record) Remove(db *database.Database) (err error) {
 	coll := db.DomainsRecords()
+
+	if r.DeleteTimestamp.IsZero() {
+		r.DeleteTimestamp = time.Now()
+
+		err = coll.CommitFields(r.Id, r, set.NewSet("delete_timestamp"))
+		if err != nil {
+			return
+		}
+
+		return
+	}
+
+	deleteTtl := time.Duration(settings.System.DomainDeleteTtl) * time.Second
+	if time.Since(r.DeleteTimestamp) < deleteTtl {
+		return
+	}
 
 	_, err = coll.DeleteOne(db, &bson.M{
 		"_id": r.Id,
